+++ src/Card/CreditCard.java
@@ -7,12 +7,12 @@ public class CreditCard extends Card {
         super(cardholder);
     }
 
-    CreditCard(String cardholder, String accountBalance) {
+    public CreditCard(String cardholder, String accountBalance) {
         super(cardholder, accountBalance);
     }
 
     @Override
-    void withdrawFromCardBalance(BigDecimal withdraw) {
+    public void withdrawFromCardBalance(BigDecimal withdraw) {
         accountBalance = accountBalance.subtract(withdraw);
     }
