+++ src/Card/exception/InsufficientFundsException.java
@@ -1,4 +1,4 @@
-package Card;
+package Card.exception;
 
 public class InsufficientFundsException extends Throwable {
     public InsufficientFundsException(String msg) {
